requirements.txt
pydantic==2.3.0
psycopg2-binary==2.9.9
Jinja2==3.1.2
PyYAML==6.0.1
alembic==1.12.0
aiogram==3.1.1
loguru==0.7.2
SQLAlchemy==2.0.25